integral over frame processor and integral over thres over frame processor inserted, now also a postprocessor exists that delivers the average over a certain time period. Now not downscaling to uint32 in saveparam function

+++ cass_pnccd/pnccd_analysis.cpp
@@ -181,7 +181,7 @@ void cass::pnCCD::Parameter::save()
     setValue("CreatePixelList",dp._createPixellist);
     setValue("DoOffsetCorrection",dp._doOffsetCorrection);
     setValue("useCommonMode",dp._useCommonMode);
-    setValue("IntegralOverThres",static_cast<int32_t>(dp._thres_for_integral));
+    setValue("IntegralOverThres",static_cast<qint64>(dp._thres_for_integral));
     setValue("DarkCalibrationFileName",dp._darkcalfilename.c_str());
     setValue("DarkCalibrationSaveFileName",dp._savedarkcalfilename.c_str());
     //cass::PixelDetector::ROIsimple::save();
